moved keymap into a module. #495
